experiments/real_world/mta_subway_hourly_tier2p11/requirements.txt
pandas>=2.2
numpy>=2.0
pyarrow>=14
PyYAML>=6
matplotlib>=3.8
